# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Autoconf Vehicles API",
        "description": "API para gerenciamento de veículos - Desafio Técnico",
        "contact": {
            "email": "contato@autoconf.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/vehicles": {
            "get": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Listar veículos com paginação e filtros",
                "operationId": "e6e93662341185561b564e4380f2e300",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página (máx. 100)",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 10
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Busca global (license_plate, brand, model)",
                        "schema": {
                            "type": "string",
                            "example": "Toyota"
                        }
                    },
                    {
                        "name": "brand",
                        "in": "query",
                        "description": "Filtrar por marca",
                        "schema": {
                            "type": "string",
                            "example": "Toyota"
                        }
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "description": "Filtrar por modelo",
                        "schema": {
                            "type": "string",
                            "example": "Corolla"
                        }
                    },
                    {
                        "name": "license_plate",
                        "in": "query",
                        "description": "Filtrar por placa",
                        "schema": {
                            "type": "string",
                            "example": "ABC1D23"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordenação (ex: km,-sale_price)",
                        "schema": {
                            "type": "string",
                            "example": "km,-sale_price"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de veículos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedVehicleResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Criar novo veículo",
                "operationId": "dc52fe989ad8732ce06f6a502154d8ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Veículo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Veículo criado com sucesso"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicles/{id}": {
            "get": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Obter detalhes do veículo",
                "operationId": "3fadcce9f8fdde1455d3e8270fd931c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do veículo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Veículo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Atualizar veículo",
                "operationId": "57a33bc2372e43b77031015215ba0c4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Veículo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Veículo atualizado com sucesso"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Ação não autorizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Veículo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Excluir veículo",
                "operationId": "6bdefb6386f49ad654b9db474c22db2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veículo excluído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Veículo excluído com sucesso"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Ação não autorizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Veículo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao excluir o veículo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VehicleStoreRequest": {
                "required": [
                    "license_plate",
                    "chassis",
                    "brand",
                    "model",
                    "sale_price",
                    "transmission",
                    "fuel_type"
                ],
                "properties": {
                    "license_plate": {
                        "type": "string",
                        "pattern": "^[A-Z]{3}[0-9][A-Z0-9][0-9]{2}$",
                        "example": "ABC1D23"
                    },
                    "chassis": {
                        "type": "string",
                        "maxLength": 17,
                        "minLength": 17,
                        "example": "9BWZZZ377VT004251"
                    },
                    "brand": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "version": {
                        "type": "string",
                        "example": "XEI 2.0"
                    },
                    "sale_price": {
                        "type": "number",
                        "format": "decimal",
                        "minimum": 0.01,
                        "example": 85000
                    },
                    "color": {
                        "type": "string",
                        "example": "Prata"
                    },
                    "km": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 15000
                    },
                    "transmission": {
                        "type": "string",
                        "enum": [
                            "manual",
                            "automatic"
                        ],
                        "example": "automatic"
                    },
                    "fuel_type": {
                        "type": "string",
                        "enum": [
                            "gasoline",
                            "ethanol",
                            "flex",
                            "diesel",
                            "hybrid",
                            "electric"
                        ],
                        "example": "flex"
                    }
                },
                "type": "object"
            },
            "PaginatedVehicleResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Veículos listados com sucesso"
                    },
                    "data": {
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            },
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 5
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 10
                            },
                            "total": {
                                "type": "integer",
                                "example": 50
                            },
                            "from": {
                                "type": "integer",
                                "example": 1
                            },
                            "to": {
                                "type": "integer",
                                "example": 10
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "license_plate": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "O campo license_plate é obrigatório."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UnauthorizedError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ForbiddenError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "NotFoundError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "No query results for model [Vehicle] 1"
                    }
                },
                "type": "object"
            },
...
            "VehicleSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Veículo encontrado com sucesso"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Vehicle"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}